refactor: move schema and repo to infrastructure folder
