+++ services/api-main/src/app/notifications/schema/notification.schema.ts
@@ -1,4 +1,4 @@
-import { Ticket } from './../../tickets/schema/ticket.schema';
+import { Ticket } from '../../tickets/infrastructure/schema/ticket.schema';
 import { Schema, Prop, SchemaFactory } from '@nestjs/mongoose';
 import mongoose, { Document } from 'mongoose';
 import { User } from 'src/app/users/schema/user.schema';
