+++ services/api-main/src/app/tickets/profiles/ticket.profile.ts
@@ -8,9 +8,9 @@ import {
   mapWithArguments,
 } from '@automapper/core';
 import { Injectable } from '@nestjs/common';
-import { Ticket } from '../schema/ticket.schema';
+import { Ticket } from '../infrastructure/schema/ticket.schema';
 import { TicketDTO } from '../dto/ticket.dto';
-import { TicketHistoryItem } from '../schema/ticket-history.schema';
+import { TicketHistoryItem } from '../infrastructure/schema/ticket-history.schema';
 import { TicketHistoryItemDTO } from '../dto/ticket-history.dto';
 import { User } from 'src/app/users/schema/user.schema';
 import { UserDTO } from 'src/app/users/dto/user.dto';
