+++ services/api-main/src/app/tickets/tickets.module.ts
@@ -2,11 +2,14 @@ import { Module } from '@nestjs/common';
 import { TicketsService } from './tickets.service';
 import { TicketsController } from './tickets.controller';
 import { MongooseModule } from '@nestjs/mongoose';
-import { Ticket, TicketSchema } from 'src/app/tickets/schema/ticket.schema';
+import {
+  Ticket,
+  TicketSchema,
+} from 'src/app/tickets/infrastructure/schema/ticket.schema';
 import { UsersModule } from 'src/app/users/users.module';
 import { TicketTagSystemModule } from '../ticket-tag-system/ticket-tag-system.module';
 import { NotificationsModule } from '../notifications/notifications.module';
-import { TicketsRepository } from './tickets.repository';
+import { TicketsRepository } from './infrastructure/tickets.repository';
 
 @Module({
   imports: [
@@ -15,6 +18,7 @@ import { TicketsRepository } from './tickets.repository';
     TicketTagSystemModule,
     NotificationsModule,
   ],
+
   controllers: [TicketsController],
   providers: [TicketsService, TicketsRepository],
 })
